@@ -7,9 +7,7 @@ import org.junit.runners.Suite;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({ // @formatter:off
-     JacksonDeserializationUnitTest.class
-    ,JacksonDeserializationUnitTest.class
-    ,JacksonPrettyPrintUnitTest.class
+     JacksonPrettyPrintUnitTest.class
     ,SandboxUnitTest.class
 }) // @formatter:on
 public class UnitTestSuite {
